detection/EyesDetection/requirements.txt
opencv-python
tensorflow
numpy
pyserial
spidev
RPi.GPIO
